+++ docs/changelog.rst
@@ -1,5 +1,10 @@
 Changelog
 =========
+Next Release
+------------
+:py:func:`vsfieldkit.prepare_nnedi3_chroma_upsampler` can now use the znedi3
+or nnedi3cl plugin in addition to plain nnedi3 as before.
+
 2.0.1
 -----
 Fix :py:func:`vsfieldkit.resample_as_progressive`'s ``avoid_chroma_shift``