+++ docs/changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
-Next Release
-------------
+2.1.0
+-----
 * :py:func:`vsfieldkit.prepare_nnedi3_chroma_upsampler` can now use the znedi3
   or nnedi3cl plugin in addition to plain nnedi3 as before.
 * Add compatibility with VapourSynth>=63 Python wrapper, broken due to renamed